# Speed up Convert.ToBase64String()