It is 30% performance optimization:
```
|                                      Method |     Mean | Error |      Min |      Max |   Gen 0 | Allocated |
|-------------------------------------------- |---------:|------:|---------:|---------:|--------:|----------:|
|                  Old_Convert_ToBase64String | 540.4 us |    NA | 540.4 us | 540.4 us |  9.0000 |      1 MB |
| Old_Convert_ToBase64String_InsertLineBreaks | 634.3 us |    NA | 634.3 us | 634.3 us | 10.0000 |      1 MB |
|                  New_Convert_ToBase64String | 396.4 us |    NA | 396.4 us | 396.4 us |  9.0000 |      1 MB |
| New_Convert_ToBase64String_InsertLineBreaks | 425.9 us |    NA | 425.9 us | 425.9 us |  9.0000 |      1 MB |
```
Benchmarked on 1000 random byte arrays of lengths 0..999

Used tricks:
* Use precomputed table of Base64 char pairs (takes 16 KiB RAM and some warmup time to initialize (once) and load to CPU Cache). It reduces number of memory operations 2 times. Use `int` to process two char values simultaneously.
* Avoid reading each byte from `inData` array twice.
* More fast pointer arithmetic:  `*ptr++`  instead of `ptrBase[index]; ++index;`. We are saving `add` instruction.
* Get rid of `offset` parameter from `ConvertToBase64Array()`. `inData` is `array base + offset`.
* Optimize most critical loop: check only one condition inside it. `insertLineBreaks` impacts on number of loop rounds before next complex condition check.
* Consider `BitConverter.IsLittleEndian` for correct work on Big-endian platforms.