@EgorBo 2022-05-30
There was a PR to use SSE for this API - https://github.com/dotnet/coreclr/pull/21833