@EgorBo 2022-05-30
> > There was a PR to use SSE for this API - [dotnet/coreclr#21833](https://github.com/dotnet/coreclr/pull/21833)
> 
> Why didn't we finish it?

It got lost during coreclr->runtime migration 🙂 I can port it to crossplat intrinsics after(if) this PR lands