@stephentoub 2022-05-31
> It got lost during coreclr->runtime migration 🙂 I can port it to crossplat intrinsics after(if) this PR lands to avoid conflicts

A quick skim of that PR suggests it doesn't require a 16K lookup table?  If that's the case, with appreciation to Sergei, I'd prefer we just start with that PR as something that's faster and smaller.