@EgorBo 2022-06-29
> @EgorBo, are you still going to look at bringing that PR back to life?

Sure, will take a look in an hour