@EgorBo 2022-06-29
> > @EgorBo, are you still going to look at bringing that PR back to life?
> 
> Sure, will take a look in an hour

So I took a look - there was a small bug in the impl, but the problem that it will be not trivial to extend it to support Arm - it has to many non-shared intrinsics, I'll try to port it in coming days