@SergeiPavlov 2022-07-22
Yes.

The function is optimized by https://github.com/dotnet/runtime/commit/a959c3ee2f4ae21e1eb3a6d8f4089ecdea76d649